Move type definitions into route.ts files
