--- client/finance/app/api/auth/check-permission/types.ts
@@ -1,16 +1,0 @@
-import { Feature, PermissionLevel } from '@/types/AuthorizationTypes';
-
-/**
- * 権限チェックAPIのリクエスト型
- */
-export interface CheckPermissionRequestType {
-  feature: Feature;
-  level: PermissionLevel;
-}
-
-/**
- * 権限チェックAPIのレスポンス型
- */
-export interface CheckPermissionResponseType {
-  hasPermission: boolean;
-}
