--- client/finance/app/api/permission-matrix/types.ts
@@ -1,22 +1,0 @@
-import { PermissionMatrix } from '@/types/AuthorizationTypes';
-
-/**
- * 権限マトリックス更新APIのリクエスト型
- */
-export interface PermissionMatrixUpdateRequestType {
-  matrix: PermissionMatrix;
-}
-
-/**
- * 権限マトリックス取得APIのレスポンス型
- */
-export interface PermissionMatrixGetResponseType {
-  matrix: PermissionMatrix;
-}
-
-/**
- * 権限マトリックス更新APIのレスポンス型
- */
-export interface PermissionMatrixUpdateResponseType {
-  message: string;
-}
