+++ client/finance/app/components/FeatureGuard.tsx
@@ -3,10 +3,10 @@
 import { useEffect, useState } from 'react';
 
 import { Feature, PermissionLevel } from '@/types/AuthorizationTypes';
-import {
+import type {
   CheckPermissionRequestType,
   CheckPermissionResponseType,
-} from '@/app/api/auth/check-permission/types';
+} from '@/app/api/auth/check-permission/route';
 
 interface FeatureGuardProps {
   feature: Feature;
